Moving custom hooks to another folder
